+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Variable temperature sizing (at least in the code version)
 - Reimplemented size by length and width
 
+### Changed
+- In figure plotting, the interval[x[i], x[i+1]) now has the value y[i] (instead of y[i-1]).
+- Scroll behaviour on the result page (issue #99).
+
+### Fixed
+- Wrong heating load in april in GUI (issue #94).
+
 ## [2.1.1] - 2023-01-30
 
 ### Added